requirements.txt
# dependencies

Flask >= 1.1, <2
gunicorn >= 20
selenium >= 3.141, <4
chromedriver-binary >= 93, <94
urllib3 >= 1.21, < 2
psycopg2-binary >= 2.8, <3
requests >= 2.18, <3
lxml >= 4.5, <5
SQLAlchemy >= 1.3, <2
google-cloud-tasks >= 1.5.0, < 2
google-cloud-storage >= 1.30.0, < 2
google-cloud-logging >= 1.15.1, < 2
google-cloud-secret-manager >= 2.7.0, < 3